(PC-29736) refactor(tabbar): Extract inner component to reuse container
