+++ src/features/navigation/TabBar/TabBarComponent.tsx
@@ -3,17 +3,11 @@ import { Platform } from 'react-native'
 import styled from 'styled-components/native'
 
 import { menu } from 'features/navigation/TabBar/menu'
-import { TabBarTitle as Title } from 'features/navigation/TabBar/TabBarTitle'
+import { TabBarInnerComponent } from 'features/navigation/TabBar/TabBarInnerComponent'
 import { TabRouteName } from 'features/navigation/TabBar/types'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
 import { InternalNavigationProps } from 'ui/components/touchableLink/types'
-import { BicolorLogo } from 'ui/svg/icons/BicolorLogo'
-import { BicolorSelector } from 'ui/svg/icons/BicolorSelector'
 import { AccessibleBicolorIcon } from 'ui/svg/icons/types'
-import { Spacer, getSpacing } from 'ui/theme'
-
-const SELECTOR_WIDTH = '80%'
-const SELECTOR_HEIGHT = getSpacing(1)
 
 interface Props {
   isSelected?: boolean
@@ -42,28 +36,9 @@ export const TabBarComponent: React.FC<Props> = ({
     accessibilityLabel={menu[tabName].accessibilityLabel}
     testID={menu[tabName].accessibilityLabel ?? menu[tabName].displayName}
     accessibilityCurrent={isSelected ? 'page' : undefined}>
-    {isSelected ? (
-      <BicolorSelector
-        width={SELECTOR_WIDTH}
-        height={SELECTOR_HEIGHT}
-        testID={`${menu[tabName].accessibilityLabel} sélectionné`}
-      />
-    ) : null}
-    <Spacer.Flex />
-    <StyledIcon as={BicolorIcon} selected={isSelected} />
-    <Title selected={isSelected} displayName={menu[tabName].displayName} />
-    <Spacer.Flex />
-    {isSelected ? <BicolorSelectorPlaceholder /> : null}
+    <TabBarInnerComponent tabName={tabName} isSelected={isSelected} BicolorIcon={BicolorIcon} />
   </TabComponentContainer>
 )
-
-const StyledIcon = styled(BicolorLogo).attrs<{ selected?: boolean }>(({ theme, selected }) => ({
-  color: selected ? undefined : theme.colors.greyDark,
-  size: theme.tabBar.iconSize,
-  thin: !selected,
-}))<{ selected?: boolean }>``
-
-const BicolorSelectorPlaceholder = styled.View({ height: SELECTOR_HEIGHT })
 
 const TabComponentContainer: typeof InternalTouchableLink = styled(InternalTouchableLink).attrs(
   ({ theme, accessibilityLabel, selected }) => ({
